Update package.json & fix index.mjs to use ESM
