+++ package.json
@@ -1,16 +1,18 @@
 {
-  "name": "hyprland-site",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
+  "name": "hyprland-website",
+  "version": "1.1.0",
+  "description": "Website for Hyprland - Hyprland - A wayland compositor that doesn't sacrifice on its looks!",
+  "main": "index.mjs",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "start": "node index.js"
   },
-  "keywords": [],
+  "keywords": ["hyprland"],
   "author": "",
-  "license": "ISC",
+  "license": "BSD-3-Clause",
   "dependencies": {
     "ejs": "^3.1.8",
     "express": "^4.18.1"
-  }
+  },
+  "repository": "github:hyprwm/hyprland-website",
+  "private": "true"
 }
